A lot of files overhauled to merge what Sam did on the backend with
Chris' work on the frontend. Check the details for changes.

+++ src/pages/functions/functions.ts
@@ -26,9 +26,9 @@ export class System {
 
 	constructor(public toastCtrl: ToastController, public alertCtrl: AlertController, public loadingCtrl: LoadingController, private movesService: MovesService) {
 
-  	}
+  }
 
-showNotification(msg, duration) {
+  showNotification(msg, duration) {
     let toast = this.toastCtrl.create({
       message: msg,
       duration: duration
@@ -45,65 +45,62 @@ startLoading(msg, duration) {
 	loader.present();
 }
 
-moveOptionsScreen(move) {
-      let confirm = this.alertCtrl.create({
-            message: 'Are you sure want to delete "' + move.info.name + '"?',
-    buttons: [
-      {
-        text: 'Cancel',
-        role: 'cancel',
-        handler: data => {
-          console.log('Cancel clicked');
-        }},
+  moveOptionsScreen(move) {
+    let confirm = this.alertCtrl.create({
+      message: 'Are you sure want to delete "' + move.info.name + '"?',
+      buttons: [
         {
-        text: 'Yes',
-        handler: data => {
-          this.deleteMove(move);
-        }}]
-        });
-          confirm.present();
+          text: 'Cancel',
+          role: 'cancel',
+          handler: data => {
+            console.log('Cancel clicked');
+          }
+        },
+        {
+          text: 'Yes',
+          handler: data => {
+            this.deleteMove(move);
+          }
         }
-
-  listMoves() {
-    this.movesService.getMoves().subscribe(
-      data => {
-        this.moves = data;
-        console.log(this.moves);
-        this.moves.sort(this.sortDescending);
-      },
-      err => {
-        console.log(err);
-      },
-      () => console.log('Got Moves')
-    );
+      ]
+    });
+    
+    confirm.present();
   }
 
+  
+  // listMoves() {
+  //   this.movesService.getMoves().then((data) => {
+
+  //     this.moves = data;
+
+  //   }, (err) => {
+
+  //     console.log(err);
+
+  //   });
+  // }
+  
+
 deleteMove(move) {
-    this.movesService.deleteMove(move).subscribe(
-      err => {
-        console.log(err);
-      }
-    )
+    this.movesService.deleteMove(move).then((result) => {
+
+      console.log("Deleted")
+
+    }, (err) => {
+
+      console.log(err);
+
+    });
+
+
     this.startLoading('Deleting move, standby...', 1000);
     setTimeout(() => {
         this.checked = 0;   
-        this.listMoves();  
         this.showNotification('Move has been deleted.', 1000);
     }, 1000);    
   }
 
-  saveMove(move) {
-    this.movesService.makeMove(move).subscribe(
-      data => {
-        // this.moves = data;
-        console.log(move);
-      },
-      err => {
-        console.log(err);
-      },
-      () => console.log('Got Moves')
-    );
-  }
